Worked on Sessions and more styling of the extension
